adds gender and speed to refresults table, fix to duration, and adds start and end times to json that is stored
Former-commit-id: bef1bf0bc62308fe59aa8b5cc334837f10c3b1f3 [formerly f0aef4aee087520add42a5f0b7673ee687c5d597] [formerly bef1bf0bc62308fe59aa8b5cc334837f10c3b1f3 [formerly f0aef4aee087520add42a5f0b7673ee687c5d597] [formerly bef1bf0bc62308fe59aa8b5cc334837f10c3b1f3 [formerly f0aef4aee087520add42a5f0b7673ee687c5d597] [formerly 454b4e977d3af6f5dbd11d9506ebc2b1e7657107]]]
Former-commit-id: 9e589bcead1fe071567c67185b5d41d81d00059a [formerly eab345ffca1790e62a5f23f9412bb968f2136dd1] [formerly 9e589bcead1fe071567c67185b5d41d81d00059a [formerly eab345ffca1790e62a5f23f9412bb968f2136dd1] [formerly 511bef1a9b61b5f9d5230670b68c89a94022b4d0]] [formerly 211a013f7d591f0f4dc2f03680b0e376cc096800 [formerly 2964b6f1aebbb669d5ac39971a4e954c88488a8c] [formerly 211a013f7d591f0f4dc2f03680b0e376cc096800 [formerly 2964b6f1aebbb669d5ac39971a4e954c88488a8c] [formerly 0e481a205b542c74818cb46e39859cea61af8cb1]]]
Former-commit-id: 211a013f7d591f0f4dc2f03680b0e376cc096800 [formerly 2964b6f1aebbb669d5ac39971a4e954c88488a8c]
Former-commit-id: ebc5d32abb45712f3095a3b003e1fefd718178ae [formerly 1da1a5ca1f99505aa01727fd865b4d95dff46d98] [formerly 010c5364b1f7137ab6a658a749300459aa22a824]
Former-commit-id: 6906348787b9d985031e2f15ec2145c8523245d9
Former-commit-id: 57c123d79f3c9e7ef5182215ab5ec1f32a956e53

+++ src/mitll/langtest/server/audio/AudioCheck.java
@@ -174,20 +174,20 @@ public class AudioCheck {
 
   public static class ValidityAndDur {
     public final AudioAnswer.Validity validity;
-    public final int durationInMillis;
+    public final long durationInMillis;
 
     public ValidityAndDur(AudioAnswer.Validity validity) {
       this(validity, 0d);
     }
 
-    public ValidityAndDur(AudioAnswer.Validity validity, int durInMillis) {
+    public ValidityAndDur(AudioAnswer.Validity validity, long durInMillis) {
       this.validity = validity;
       this.durationInMillis = durInMillis;
     }
 
     public ValidityAndDur(AudioAnswer.Validity validity, double dur) {
       this.validity = validity;
-      this.durationInMillis = (int) (1000d * dur);
+      this.durationInMillis = (long) (1000d * dur);
     }
     public String toString() { return "valid " + validity + " dur " + durationInMillis; }
   }
